// client/src/app/interfaces/customers-addr-item.ts
export interface CustomersAddr {
  ID: number;
  ADDRESS_TYPE: string;
  CUSTOMERS_NUMBER: string;
  ADDRESS_ISO_CODE: string;
  ADDRESS_CRYNAME: string;
  ADDRESS_STREET: string;
  ADDRESS_CITY: string;
  ADDRESS_POSTCODE: string;
  ADDRESS_COMMENT: string;
  TAXCODE: string;
  NAME_ADDR: string;
  EMAIL: string;
  PHONE: string;
  ADDRESS_ID: string;
}